# Using another set nix github actions with the idea to speed up CI