--- a/.github/workflows/integration.yml
+++ b/.github/workflows/integration.yml
@@ -23,8 +23,10 @@ jobs:
       run:
         working-directory: ./daemon
     steps:
-        - uses: actions/checkout@v3
+        - uses: actions/checkout@v4
         - name: Install Nix
-          uses: cachix/install-nix-action@v17
+          uses: DeterminateSystems/nix-installer-action@main
+        - name: Install Magic Nix Cache
+          uses: DeterminateSystems/magic-nix-cache-action@main
         - name: Run fuzzer
           run: nix develop --command cargo run --bin=fuzzer