@blinry 2024-07-16
I like this! :) Reducing energy use a bit seems good, and the change is indeed minimal.

We should cherry-pick the unrelated changes to main.